consolidated getUserData() to extrenal file
