+++ src/js/common/components/RequiredLoginView.jsx
@@ -3,22 +3,23 @@ import { connect } from 'react-redux';
 
 import { commonStore } from '../store';
 
-import environment from '../helpers/environment.js';
+import getUserData from '../helpers/login-helpers';
+import handleVerify from '../helpers/verify-user';
+
 import { updateLoggedInStatus, updateLogInUrl, updateProfileField } from '../actions';
 
 class RequiredLoginView extends React.Component {
   constructor(props) {
     super(props);
     this.handleLogin = this.handleLogin.bind(this);
-    this.handleVerify = this.handleVerify.bind(this);
+    this.handleVerify = handleVerify();
     this.setMyToken = this.setMyToken.bind(this);
-    this.getUserData = this.getUserData.bind(this);
   }
 
   componentDidMount() {
     if (localStorage.userToken) {
       this.props.onUpdateLoggedInStatus(true);
-      this.getUserData();
+      getUserData();
     } else {
       this.props.onUpdateLoggedInStatus(false);
     }
@@ -28,33 +29,8 @@ class RequiredLoginView extends React.Component {
 
   setMyToken() {
     if (event.data === localStorage.userToken) {
-      this.getUserData();
+      getUserData();
     }
-  }
-
-  getUserData() {
-    fetch(`${environment.API_URL}/v0/user`, {
-      method: 'GET',
-      headers: new Headers({
-        Authorization: `Token token=${localStorage.userToken}`
-      })
-    }).then(response => {
-      return response.json();
-    }).then(json => {
-      const userData = json.data.attributes.profile;
-      if (userData.loa.current === 1 && userData.loa.highest === 3) {
-        this.handleVerify();
-      } else {
-        this.props.onUpdateProfile('accountType', userData.loa.current);
-        this.props.onUpdateProfile('email', userData.email);
-        this.props.onUpdateProfile('userFullName.first', userData.first_name);
-        this.props.onUpdateProfile('userFullName.middle', userData.middle_name);
-        this.props.onUpdateProfile('userFullName.last', userData.last_name);
-        this.props.onUpdateProfile('gender', userData.gender);
-        this.props.onUpdateProfile('dob', userData.birth_date);
-        this.props.onUpdateLoggedInStatus(true);
-      }
-    });
   }
 
   handleLogin() {
@@ -63,14 +39,6 @@ class RequiredLoginView extends React.Component {
     const myLoginUrl = login.loginUrl.first;
     // TODO(crew): Check on how this opens on mobile.
     const receiver = window.open(myLoginUrl, '_blank', 'resizable=yes,top=50,left=500,width=500,height=750');
-    receiver.focus();
-  }
-
-  handleVerify() {
-    const myStore = commonStore.getState();
-    const login = myStore.login;
-    const myVerifyUrl = login.loginUrl.third;
-    const receiver = window.open(myVerifyUrl, '_blank', 'resizable=yes,top=50,left=500,width=500,height=750');
     receiver.focus();
   }
